DRY the code a bit
